requirements.txt
asgiref==3.6.0
asyncio==3.4.3
click==8.1.3
Django==4.1.6
django-tinymce==3.5.0
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.4.0
pip==20.0.2
psycopg2-binary==2.9.5
setuptools==44.0.0
sqlparse==0.4.3
tzdata
Werkzeug==2.2.2
wheel==0.34.2
zipp==3.12.1